requirements.txt
ultralytics==8.0.161
opencv-python==4.9.0.80
cloudinary==1.37.0
twilio==8.11.0
python-dotenv==1.0.1
pillow==10.2.0
numpy>=1.24
openimages==0.1.5
